ML_week_04/ml_bench/requirements.txt
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
shap>=0.45.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.66.0
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
joblib>=1.3.0
